A potential proof for hasEdge.
My understanding of algebraic edges are that connections are either:

1. Explicitly, and directly, made between two vertices; or
2. The result of connecting two graphs.

Overlays don't make connections, just combine graphs on similar
vertices. Connections connect all nodes in the left graph to all nodes in the right graph.

We build two proof objects for these two cases.

The first searches for direct connections. The other searches for a
connect constructor, and then looks left for the starting vertex, then
looks right for the terminating vertex.